+++ web-server/src/models/dto/role.rs
@@ -78,8 +78,8 @@ pub struct QueryRoleDTO {
     pub disable_flag: Option<String>,
 
     #[validate(range(min = 1, message = "页码最小为1"))]
-    pub page_no: Option<u64>,
+    pub page_no: Option<usize>,
 
     #[validate(range(min = 1, message = "分页最小为1"))]
-    pub page_size: Option<u64>,
+    pub page_size: Option<usize>,
 }
